initial backend config
welcome to this commit and new branch of course, so i tried to implement something in the backend, so we atleast have that...
u need a .env file with the api key from google, with the name GOOGLE_MAPS_API_KEY, and then i created a new service, which calls the google api endpoint and returns the result as. a distance matrix, right now it has some hard coded values and is triggered via a request to the endpoint "http://localhost:5001/api/distance"

and it should return a distance matrix.

have a look and try it out kings